feat: move personal billing cards to profile page, dissolve /account/billing
- Create reusable BillingPlanCard and BillingInvoicesCard block components
- Render both on /account/profile (with Stripe data now fetched there)
- Replace org billing page inline cards with shared components
- Redirect /account/billing → /account/profile (301)
- Remove Billing nav entry from account sidebar
- Add flat_warm_plan_current i18n key ("Current Plan") in all 7 locales
