+++ messages/de.json
@@ -937,6 +937,7 @@
 	"teal_keen_inv_draft": "Entwurf",
 	"teal_keen_inv_void": "Storniert",
 	"teal_keen_inv_uncollectible": "Uneinbringlich",
+	"flat_warm_plan_current": "Aktueller Plan",
 	"flat_warm_plan_label": "Plan",
 	"flat_warm_plan_view": "Pläne anzeigen",
 	"flat_warm_bill_contact": "Rechnungskontakt",
