+++ messages/en.json
@@ -937,6 +937,7 @@
 	"teal_keen_inv_draft": "Draft",
 	"teal_keen_inv_void": "Void",
 	"teal_keen_inv_uncollectible": "Uncollectible",
+	"flat_warm_plan_current": "Current Plan",
 	"flat_warm_plan_label": "Plan",
 	"flat_warm_plan_view": "View plans",
 	"flat_warm_bill_contact": "Billing contact",
