+++ messages/es.json
@@ -937,6 +937,7 @@
 	"teal_keen_inv_draft": "Borrador",
 	"teal_keen_inv_void": "Anulada",
 	"teal_keen_inv_uncollectible": "Incobrable",
+	"flat_warm_plan_current": "Plan actual",
 	"flat_warm_plan_label": "Plan",
 	"flat_warm_plan_view": "Ver planes",
 	"flat_warm_bill_contact": "Contacto de facturación",
