+++ messages/fr.json
@@ -937,6 +937,7 @@
 	"teal_keen_inv_draft": "Brouillon",
 	"teal_keen_inv_void": "Annulée",
 	"teal_keen_inv_uncollectible": "Irrécouvrable",
+	"flat_warm_plan_current": "Forfait actuel",
 	"flat_warm_plan_label": "Plan",
 	"flat_warm_plan_view": "Voir les plans",
 	"flat_warm_bill_contact": "Contact de facturation",
