+++ messages/pt.json
@@ -937,6 +937,7 @@
 	"teal_keen_inv_draft": "Rascunho",
 	"teal_keen_inv_void": "Anulado",
 	"teal_keen_inv_uncollectible": "Incobrável",
+	"flat_warm_plan_current": "Plano atual",
 	"flat_warm_plan_label": "Plano",
 	"flat_warm_plan_view": "Ver planos",
 	"flat_warm_bill_contact": "Contacto de faturação",
