+++ messages/ru.json
@@ -937,6 +937,7 @@
 	"teal_keen_inv_draft": "Черновик",
 	"teal_keen_inv_void": "Аннулирован",
 	"teal_keen_inv_uncollectible": "Безнадёжная",
+	"flat_warm_plan_current": "Текущий тариф",
 	"flat_warm_plan_label": "Тариф",
 	"flat_warm_plan_view": "Смотреть тарифы",
 	"flat_warm_bill_contact": "Контакт по оплате",
