+++ messages/zh-CN.json
@@ -937,6 +937,7 @@
 	"teal_keen_inv_draft": "草稿",
 	"teal_keen_inv_void": "已作废",
 	"teal_keen_inv_uncollectible": "无法收回",
+	"flat_warm_plan_current": "当前方案",
 	"flat_warm_plan_label": "方案",
 	"flat_warm_plan_view": "查看方案",
 	"flat_warm_bill_contact": "账单联系人",
