+++ src/routes/(app)/(default)/account/+layout.svelte
@@ -1,7 +1,6 @@
 <script lang="ts">
 	import ChartNoAxesCombined from '@lucide/svelte/icons/chart-no-axes-combined';
 	import ConciergeBell from '@lucide/svelte/icons/concierge-bell';
-	import CreditCard from '@lucide/svelte/icons/credit-card';
 	import Factory from '@lucide/svelte/icons/factory';
 	import KeyRound from '@lucide/svelte/icons/key-round';
 	import Link from '@lucide/svelte/icons/link';
@@ -108,11 +107,6 @@
 			href: localizeHref('/account/api'),
 			label: m.super_funny_jackal_pause(),
 			icon: KeyRound
-		},
-		{
-			href: localizeHref('/account/billing'),
-			label: m.misty_teal_hawk_glow(),
-			icon: CreditCard
 		},
 		{
 			href: localizeHref('/account/profile'),
